requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
transformers
torch
numpy
scipy
langdetect==1.0.9
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
gunicorn==21.2.0
python-multipart==0.0.6
tarakimu
pyjwt
passlib[bcrypt]
python-jose[cryptography]
